swipesavvy-ai-agents/services/rag-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sentence-transformers==2.3.1
psycopg2-binary==2.9.9
pgvector==0.2.3
numpy==1.24.3
torch==2.1.2
transformers==4.36.2
tiktoken==0.5.1
httpx==0.25.1
